refactor(ui): migrate LearnMore, Profile, SignUp, and Tracker styles into pages stylesheet structure
