+++ frontend/src/styles/pages/home.css
@@ -1,0 +1,289 @@
+/**
+ * EarlyBloom home page styling.
+ *
+ * This file contains page-specific presentation for:
+ * - the landing hero
+ * - mission and feature sections
+ * - the future tools callout
+ *
+ * It intentionally keeps layout aligned with the shared app shell and
+ * shared surface system without re-defining global primitives.
+ */
+
+.home-page {
+  width: 100%;
+}
+
+.home-section {
+  padding-top: var(--space-6);
+  padding-bottom: var(--space-6);
+}
+
+.home-section--callout {
+  padding-bottom: var(--space-10);
+}
+
+/* Hero */
+
+.home-hero {
+  padding-top: var(--space-6);
+  padding-bottom: var(--space-6);
+}
+
+.home-hero__surface {
+  display: grid;
+  gap: var(--space-8);
+  padding: var(--space-7);
+  border-radius: var(--radius-xl);
+  border: 1px solid rgba(124, 92, 163, 0.12);
+  background:
+    linear-gradient(
+      180deg,
+      rgba(255, 255, 255, 0.94) 0%,
+      rgba(255, 248, 252, 0.9) 100%
+    );
+  box-shadow: var(--shadow-strong);
+  overflow: hidden;
+}
+
+.home-hero__content {
+  display: grid;
+  gap: var(--space-5);
+  min-width: 0;
+}
+
+.home-hero__eyebrow {
+  display: inline-flex;
+  align-items: center;
+  width: fit-content;
+  padding: 0.65rem 0.9rem;
+  border-radius: var(--radius-pill);
+  background: rgba(245, 239, 255, 0.92);
+  border: 1px solid rgba(111, 84, 201, 0.14);
+  color: var(--color-primary-deep);
+  font-size: 0.84rem;
+  font-weight: 800;
+  letter-spacing: 0.01em;
+}
+
+.home-hero__title {
+  margin: 0;
+  max-width: 12ch;
+  color: var(--color-heading);
+  font-size: clamp(2.2rem, 9vw, 4.8rem);
+  line-height: 0.95;
+  letter-spacing: -0.06em;
+}
+
+.home-hero__description {
+  margin: 0;
+  max-width: 60ch;
+  color: var(--color-text-soft);
+  font-size: 1rem;
+  line-height: 1.75;
+}
+
+.home-hero__actions {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+
+.home-hero__highlights {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.65rem;
+}
+
+.home-highlight-chip {
+  display: inline-flex;
+  align-items: center;
+  min-height: 38px;
+  padding: 0.65rem 0.8rem;
+  border-radius: var(--radius-pill);
+  background: rgba(255, 255, 255, 0.86);
+  border: 1px solid var(--color-border);
+  color: var(--color-text-soft);
+  font-size: 0.88rem;
+  font-weight: 700;
+}
+
+.home-hero__media {
+  display: grid;
+}
+
+.home-hero__media-card {
+  position: relative;
+  display: grid;
+  place-items: center;
+  min-height: 280px;
+  padding: var(--space-6);
+  border-radius: var(--radius-xl);
+  border: 1px solid rgba(124, 92, 163, 0.12);
+  background:
+    linear-gradient(
+      135deg,
+      rgba(217, 198, 255, 0.22) 0%,
+      rgba(255, 206, 223, 0.18) 32%,
+      rgba(201, 242, 199, 0.18) 68%,
+      rgba(204, 233, 255, 0.2) 100%
+    );
+  overflow: hidden;
+}
+
+.home-hero__image {
+  width: min(100%, 360px);
+  margin-inline: auto;
+}
+
+/* Mission */
+
+.home-mission-card {
+  background:
+    linear-gradient(
+      180deg,
+      rgba(255, 255, 255, 0.92) 0%,
+      rgba(255, 247, 252, 0.88) 100%
+    );
+}
+
+/* Section heading */
+
+.home-section-heading {
+  display: grid;
+  gap: var(--space-3);
+  margin-bottom: var(--space-7);
+}
+
+/* Features */
+
+.home-feature-grid {
+  display: grid;
+  gap: var(--space-5);
+}
+
+.home-feature-card {
+  display: grid;
+  gap: var(--space-4);
+  padding: var(--space-7);
+  border-radius: var(--radius-lg);
+  background: rgba(255, 255, 255, 0.86);
+  border: 1px solid rgba(124, 92, 163, 0.12);
+  box-shadow: var(--shadow-card);
+}
+
+.home-feature-card__number {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 3rem;
+  height: 3rem;
+  border-radius: 16px;
+  background: var(--gradient-brand);
+  color: var(--color-heading);
+  font-size: 0.98rem;
+  font-weight: 800;
+}
+
+.home-feature-card__title {
+  margin: 0;
+  color: var(--color-heading);
+  font-size: 1.12rem;
+  font-weight: 800;
+  letter-spacing: -0.03em;
+}
+
+.home-feature-card__copy {
+  margin: 0;
+  color: var(--color-text-soft);
+  line-height: 1.72;
+}
+
+/* Callout */
+
+.home-callout-card {
+  display: grid;
+  gap: var(--space-7);
+  padding: var(--space-7);
+  border-radius: var(--radius-xl);
+  background:
+    linear-gradient(
+      180deg,
+      rgba(255, 255, 255, 0.92) 0%,
+      rgba(255, 247, 252, 0.9) 100%
+    );
+  border: 1px solid rgba(124, 92, 163, 0.12);
+  box-shadow: var(--shadow-strong);
+}
+
+.home-callout-card__media {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.home-callout-card__image {
+  width: min(100%, 320px);
+  margin-inline: auto;
+}
+
+.home-callout-card__content {
+  display: grid;
+  gap: var(--space-4);
+}
+
+.home-callout-list {
+  margin: 0;
+  padding-left: 1.2rem;
+  color: var(--color-text-soft);
+  line-height: 1.8;
+}
+
+.home-callout-card__actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--space-3);
+}
+
+/* Responsive */
+
+@media (min-width: 640px) {
+  .home-hero__actions {
+    flex-direction: row;
+    flex-wrap: wrap;
+  }
+
+  .home-feature-grid {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+  }
+}
+
+@media (min-width: 900px) {
+  .home-hero {
+    padding-top: var(--space-8);
+    padding-bottom: var(--space-8);
+  }
+
+  .home-hero__surface {
+    grid-template-columns: minmax(0, 1.1fr) minmax(320px, 0.9fr);
+    align-items: center;
+    gap: var(--space-10);
+    padding: var(--space-8);
+  }
+
+  .home-feature-grid {
+    grid-template-columns: repeat(3, minmax(0, 1fr));
+  }
+
+  .home-callout-card {
+    grid-template-columns: minmax(280px, 0.9fr) minmax(0, 1.1fr);
+    align-items: center;
+  }
+}
+
+@media (min-width: 1100px) {
+  .home-hero__surface,
+  .home-callout-card {
+    padding: var(--space-9);
+  }
+}
