+++ frontend/src/styles/pages/learnmore.css
